Update gedcomParser.py
Added check for if birth is after death and if marriage dates are before spouses birthdays.

On the expected output on the course page it showed errors after the table, not in line with the parsing. Do we want to make these checks functions and then iterate through the individuals/families and call the functions for each test or do them while parsing the file?